// csfre/frontend/src/app/models/models.ts
export interface Item{
    name: string
    email:string
    phone: string
    title:string
    description:string
    image: Blob
}


export interface PostingResult {
    postingId: string
    postingDate: string
    name: string
    email: string
    phone: string
    title: string
    description: string
    image: string
}